dracut: Add an ignition-fetch.service
This invokes the new `fetch` stage which does just that - fetch
the Ignition config into `/run/ignition.json` and nothing else.

Prep for adding support for redeploying the rootfs, which is
an expensive process, so we only want to do it if the fetched
Ignition config is replacing the `rootfs`.