Add `ignition-subsequent.target` for non-Ignition boots
This is needed for Fedora CoreOS work on reprovisioning
the rootfs.  We tried to split out a systemd unit to handle
mounting the root in this commit:
https://github.com/coreos/fedora-coreos-config/pull/187/commits/a6fe3749262cad8d35d73efb657dea1c9add7dd4
with `ignition-ostree-mount-sysroot.service` but the problem
is that needs to have different dependencies/ordering if we're
doing the "Ignition boot" versus a "subsequent" boot.

Add a new target that's queued up if we detect we're doing
a "subsequent" boot.