blkid: Explicitly use C int variable for numParts
This was an issue I encountered while testing partition creation on s390x, where the
number of partitions returned by numParts was a bogus number. At first I thought it was
an endianness issue with blkid, but turned out that the C compiler's int variable size and
the Go compiler's int variable sizes are different (on s390x at least) which caused this bug.

The fix is to make numParts a C int variable to be consistent with what is returned by blkid