providers/qemu: support Ignition block device on s390x and ppc64le
Add experimental support for fetching Ignition configs via a virtio
block device with serial `ignition`.

The main advantage of this is that it is cross-platform. But for now, we
only use it on platforms which don't support the QEMU firmware config
approach, which are (of those we care about) s390x and ppc64le. We may
end up using it across the remaining platforms.

See related discussions in:
https://github.com/coreos/ignition/issues/928