add coreos-teardown-initramfs-network.service
This is a forward port of coreos-teardown-initramfs-network.service
from the spec2x branch [1] (used for RHEL CoreOS). When moving to NM
in the initrd [2] we decided that we also needed a mechanism to take down
the networking between the initramfs and the real root. While we would
like to use NetworkManager's logic to do this operation in the future
it's currently not easily achieved because NetworkManager is not running
persistently in the initramfs [3].

[1] https://github.com/coreos/ignition-dracut/pull/78
[2] https://github.com/coreos/fedora-coreos-tracker/issues/394
[3] https://bugzilla.redhat.com/show_bug.cgi?id=1814038