coreos-teardown-initramfs-network: simplify networking takedown
On recommendation from the NM team let's try to delete the device
first and if that doesn't work then set it to down and flush any
associated addresses. Deleting virtual devices (bonds, teams, bridges,
ip-tunnels, etc) will clean up any associated kernel resources, so
this makes the logic for bringing down these devices easier. A real
device can't be deleted so that will fail and we'll fallback to setting
it down and flushing addresses.