Declare external applications as optional
The Ignition configuration supports quite a few features, from setting the
user password over adding systemd services to creating partitions. For a lot of
these features external helper applications are called. However not all of
these configuration options are available or make sense on all systems - the
current code for example lists mkfs.btrfs as an optional dependency, but
mkfs.ext4 is declared as mandatory.

This is not consistent, so only add those dependencies as mandatory which are
required for Ignition to start / process an empty configuration file. All the
other applications are declared as optional.

If the Ignition configuration is using a feature where the helper application
is not installed in the initrd it will boil out with an error.