Add new ignition-fetch-offline.service
Make use of the new `fetch-offline` stage:

https://github.com/coreos/ignition/issues/979

We run this between the `setup` and `fetch` stages (the latter possibly
being skipped if networking is not required).

We hit the same issue here that `coreos-copy-firstboot-network.service`
hit, which is that we can't run before the `cmdline` hook because that
runs *before* udev, but we want the `by-*` symlinks for
`ignition-setup-user.service`.

The hack we do here is to rerun the NM cmdline hook in case ignition
dropped a snippet in `/etc/cmdline.d`. As mentioned in
https://github.com/coreos/fedora-coreos-config/pull/346, we'll be able
to do this more cleanly once we run NM as a systemd service directly.