coreos-teardown-initramfs-network: support coreos.no_persist_ip karg
Add coreos.no_persist_ip to allow for explicitly disabling the network
config propagation. This is supported on the spec2x branch of
ignition-dracut (i.e. for RHCOS) right now so we might as well have
it in the master branch so we can keep the maintenance burden lower.

Note that this adds two new functions load_dracut_libs() and
dracut_func(). We needed to create these new functions in order to
wrap in set +/- euo pipefail because dracut was written with set -eu
in mind and there are unset variables and commands that fail in a
lot of places.