coreos-teardown-initramfs-network: also propagate hostname
This is a forward port of 1f8184e (#156). If the admin specified
a hostname in the ip= karg static networking config and no hostname
was specified via Ignition then we'll make sure the hostname from
the ip= karg gets applied to the real root (persistently).

Ideally in the future there will be better support for this in
NetworkManager itself as the `network-legacy` dracut module did
at least provide more support for setting the hostname via ip=
kargs than `network-manager` currently does. The discussion
about this problem is in [1]. The fix for that will most likely
implicate changes to the propagation code introduced here.

Fixes: https://github.com/coreos/fedora-coreos-tracker/issues/466

[1] https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/issues/419