ci: run blackbox tests in CoreOS CI
Right now, the blackbox test is running in a job on the CL Jenkins
instance. We want to migrate this to CoreOS CI where we have the right
perms set up for everyone to see.

Use the new external host tests support in kola to just run it directly
in the just-built FCOS where it'll have all the privs it needs to e.g.
set up loopback devices.

Aside: I'm thinking we should be able to migrate all of those blackbox
tests to external host tests if we gained the ability to specify the
Ignition config. Booting VMs would be more expensive of course, but way
closer to how Ignition is actually run and it would allow cleaning up
the blackbox testing-specific code. We could also do tricks like not
actually switch root to make each test less expensive.