30ignition: drop hard requirement on qemu_fw_cfg
The original intent here was that on arches where `fw_cfg` is available,
then we definitely want to make sure that either it's built into the
kernel or if a module, added to the initrd. However, this check has been
watered down now to to the point where it's not that useful.

In https://github.com/coreos/ignition-dracut/issues/176, there's a
reasonable request to support not having the `qemu_fw_cfg` module
installed at all even if the architecture would support it.

IMO, I think we should just drop this whole logic entirely. The `qemu`
module (which we depend on) already does an `instmods qemu_fw_cfg`
(without `-c`), and qemu testing is the backbone of our CI testing. So
realistically, we would notice very fast if `fw_cfg` support in the
initrd went missing.

Closes: #176