doc/development: update spec stabilization instructions
We don't need to update the stable config version in the new experimental
config.go, since individual specs no longer delegate to the previous spec
if a config version is unrecognized.  For the same reason, the new
experimental spec should test that the previous stable version is invalid.

We now autogenerate blackbox tests per spec version, so we only need to
update the VersionOnlyConfig negative test.

Document docs updates.

Other clarifications and fixes.