config/*: drop warning on unset file/dir mode
The warning was originally added in 424f37b75583 because Ignition spec 2
defaulted to mode 0000, which was certainly not what the user wanted.
However, spec 3 defaults to 0755 for dirs and 0644 for files, and is
documented as such.  It's therefore reasonable for the user to omit those
fields, and by issuing the warnings we're encouraging the user to write
unnecessarily verbose Ignition configs (or FCCs).