stages/disks: warn if libblkid finds multiple filesystem signatures
When checking whether an existing filesystem matches the config, detect
multiple filesystem signatures on a target partition.  If found, log a
warning.  Don't do more than that for now, for backward compatibility
and because it's not clear that we should reject an otherwise-matching
filesystem just because it contains some leftover bits.

Disallow multiple signatures during test output validation, unless the
test requests otherwise.