Azure: allow for either udf or iso9660 config volume
Azure and it's varients use a cdrom for delivering the configuration
data. On Azure proper, the volume type is a udf. On Azure varient the
device may be either an iso9660 or a udf volume.

The previous behavior of the Azure provider was to attempt to mount a udf
volume. In order to support the variant Azures (e.g. Azure Stack),
Ignition will check the filesystem found on the cdrom to make sure its a
known valid type.

Signed-off-by: Ben Howard <ben.howard@redhat.com>