util/user_group_lookup: handle more error codes
Both `getpwnam_r(3)` and `getgrnam_r(3)` can return `ENOENT` or `ESRCH`
if the record is not found. Handle these.

The manpages also list `EBADF` and `EPERM` (and implies others via
`...`) as other possible codes, which is not great. Digging into this a
bit, I think it may be due to different conventions in different NSS
providers? Let's just handle `ENOENT` and `ESRCH` for now and leave the
less obvious ones until we actually hit something that returns it.

This fixes the error CI is hitting in #1014 where it gets `ESRCH` when
looking up a non-existent group.

Tested-by: Sohan Kunkerkar <sohank2602@gmail.com>