files: don't relabel homedir symlinks themselves
Regression from #996. If the home directory is a symlink, then just
relabel the referent, not the symlink itself. Since the symlink already
existed, we assume that it's properly labeled.

This causes an error on the FCOS live ISO, where `/sysroot` is mounted
from the squashfs, and so is read-only. But even on non-live, we should
just assume that whatever created the `/root -> /var/roothome` symlink
labeled it correctly.

This would normally be a no-op because `setfiles` would see that it's
properly labeled and not even attempt a `setxattr`. But because we can't
yet read SELinux labels from the initrd, it thinks it's unlabeled. (That
will be fixed by https://bugzilla.redhat.com/show_bug.cgi?id=1845210).