types: Use "path" instead of "path/filepath"
The functions from the "path" package are idempotent while the
implementations of functions in "path/filepath" differ between unix and
windows systems.

This is so the Link validation as well as the Node.Depth() function work
correctly when run on windows systems, as Ignition config should
always be validated against unix-style paths, and never
against windows-style paths. This commit ensures that.