Write SSH keys directly to authorized_keys, not to fragment file
The MCO currently expects to manage SSH keys directly in authorized_keys.
As a result, any keys written to the fragment file can't be removed by
the MCO.  To avoid this, we'll stop RHCOS from reading the fragment file,
but that would make Ignition-written keys ineffective unless we also
have Ignition write directly to authorized_keys.

https://bugzilla.redhat.com/show_bug.cgi?id=1868062