fetch: don't run if fetch-offline fetched a config
It's possible that `fetch-offline` does successfully fetch and cache a
config but still touch `/run/ignition/neednet`. That stamp file is an
API between Ignition and the OS that Ignition will need networking to
fully process it. But the config itself might've successfully been
fetched (and the only thing that requires networking is e.g. a remote
file). In those cases, there's no point in running `fetch` and have it
query the provider again.