Move all AWS session set to Fetcher
Currently, fetching configs from S3 works if the file is in a region
other than us-east-1, but fetching "files" content from S3 does not.

Rather than figuring out the region based on the session in multiple
places, just do it all as part of the session setup in the S3 URL
Fetcher.