fetch-offline: immediately return ErrNeedNet on OpenStack
This is a follow up to (and revert of) #1057.

On OpenStack, we don't actually know if we're fetching from the
config-drive or from the metadata server.

In theory, if it's from the config-drive, we don't strictly need
networking, and so `fetch-offline` would work. The issue is that in the
more common case of the metadata server, `fetch-offline` will still wait
the full 30s for the config-drive to also time out only to then have to
bring up networking and run the `fetch` stage.

Instead, let's just accept the brokenness of the OpenStack provider and
declare it as always requiring networking.

For more information, see:
https://github.com/coreos/ignition/issues/1081#issuecomment-693022160