config/merge: allow obtaining a transcript of merge operations
If FCCT wants to use config merging to implement desugaring, it needs
to be able to map fields in the merged config back to the original FCC
YAML, which means that it needs a mapping from input to output fields.
Add a merge function that also returns a transcript of field mappings.