ci: add temporary workaround for external tests using experimental specs
Instead of temporarily disabling tests, add a temporary workaround in
the Ignition CI with a built-in time bomb to make sure we don't forget
to eventually remove it once it's no longer needed. This will allow us
to stay covered by CI at least on the fedora-coreos-config side
throughout the whole stabilization process.

Ideally, we stay covered in this repo as well, but we can't do that yet
because kola still wants to parse the config.