tools/docs: notice invalid configs with stable specs
In Ignition 2.x, the Parse() function from an individual config package
doesn't chain to earlier specs, but this code still assumed it does.
As a result, all configs with stable spec versions would fail on
ErrUnknownVersion, but we wouldn't notice because we were only looking
at the report and not the error return.

Switch to the high-level chaining parse function.  Also, check for parse
errors that aren't included in the report.