main: drop -clear-cache flag
It was added in 0dcf378e2b7c, at a time when any stage would refetch the
config if the cache was unavailable (see also 9b2913044a05).  The cache
is now populated by dedicated fetch stages and later stages will fail if
it's missing, so we can no longer clear it at will.

I'm not aware of any code using this flag.  Anything that needs it is
running Ignition more than once per boot (i.e. test code), and should
specify -config-cache directly (either with a distinct file for each run,
or a file in a known path that it can remove between runs).