providers/qemu: optimize fw_cfg read size
The kernel returns a maximum of 4096 bytes at a time, and firmware reads
take O(offset) time, so we should try to optimize our read syscalls.
ioutil.ReadFile() eventually converges to an I/O size >= 4096 bytes but
we can do it immediately and without repeatedly reallocating the buffer.
This saves 8 read operations at the beginning of the config, which is a
modest but non-zero improvement.