providers/virtualbox: read config from /Ignition/Config guest property
VirtualBox supports "guest properties", which are persistent UTF-8
key-value pairs accessible from both the host and guest.  Property values
were originally limited to 128 bytes, and the VirtualBox codebase still
contains vestiges of that limit, such as fixed-size buffers in some client
code.  The limit is no longer enforced by the VirtualBox core; it seems to
have been removed in r21629 in 2009.

Read Ignition configs from an /Ignition/Config guest property by talking
directly to the upstream vboxguest kernel module via /dev/vboxguest.
Ignition configs can be attached to a VM with something like this command:

    VBoxManage guestproperty set vm-name /Ignition/Config "$(cat config.ign)"

When using this approach (rather than the VirtualBox XPCOM API), configs
are subject to these limits:

    Linux: 128 KiB (MAX_ARG_STRLEN)
    macOS: 256 KiB (ARG_MAX)
    Windows bash: 32 KiB (UNICODE_STRING.MaximumLength)
    Windows cmd.exe or PowerShell: 8 KiB

(The limits are actually slightly lower for various reasons.)

For forward compatibility, check that the /Ignition/Config/Encoding
property is missing or empty.  In the future we might want to support
alternate encodings (e.g. gzip+base64), similar to the VMware provider.

Drop support for the old magic partition GUID.  It was implemented for
the old Container Linux Vagrant provider, which was never widely deployed.
It isn't practical for hand-crafting a config drive, since it requires
creating a disk image with a partition table and then inserting the config
into a partition, and I'm not aware of any active users in the wild.

Fixes https://github.com/coreos/ignition/issues/629.