tests/filesystem: fix umountPartition retry loop
This originally called umount(8), and on failure, checked to see if the
unmount had succeeded anyway.  A later refactor switched to umount(2) but
failed to update the error message, and also broke the retry loop so that
it no longer ignored a spurious error return if umount actually succeeded.

Reinstate the success return if the filesystem was umounted, regardless of
the umount(2) return value.  Also clarify our error message on failure and
include the error we got from the syscall.

Fixes: 634508423b63 ("tests/filesystems: fix error handling")