workflows: build and test ignition-validate on macOS and Windows
We ship ignition-validate for macOS and Windows, so let's test that in
CI.  Ignition contains a lot of Linux-specific code, so it's not practical
to run through the entire ./build and ./test flow.  Instead, create a
parallel job that builds ignition-validate and runs only those tests that
are pertinent to it.