Add ignition-delete-config.service and ignition-rmcfg symlink
Add a systemd service that runs ignition-rmcfg to delete userdata early
in first boot on VirtualBox and VMware.  This will fail the boot if it
runs unsuccessfully, and can be masked in the config if desired.  Add
/usr/libexec/ignition-rmcfg as a symlink to the Ignition binary in the
Dracut module, to save space.

Distros will probably want to enable this service by default.

Fixes https://github.com/coreos/ignition/issues/1315.