internal/resource: derive AWS region hint from ARN partition field
For S3 access point ARNs, we derive the access point region from the
region field in the ARN, allowing us to query the correct AWS partition.
For regular S3 object ARNs, the region field is not used, so we assume
the same partition Ignition is running in, or "aws" if running outside
AWS.

However, object ARNs include a valid partition field.  Compute a region
hint from that, and use it to ask the partition for the correct bucket
region.  That allows object ARNs to reference objects outside the current
partition (when running in AWS) or outside the "aws" partition (when
running outside AWS).