resource: don't create our own logger
It's possible and normal to get to fetchFromHTTP() without having an
f.client, but as far as I can tell, there are no code paths that get there
without having an f.Logger.  Don't override the one that's already there.

This code was added in 9526d0fa and commented upon in
https://github.com/coreos/ignition/pull/566#discussion_r198947266, but
it appears that there was no followup at the time.

Fixes https://github.com/coreos/ignition/issues/1421.