stages/disks: fail when attempting to reuse a LUKS1 volume
Ignition has never supported LUKS1, so we can't ever try to reuse a
LUKS1 volume created by Ignition.  However, a LUKS1 volume may have been
created outside of Ignition, in which case we would conclude that the
volume was not LUKS and clobber it, causing data loss.  Detect this case
and fail instead.