Initial support for automatic spec doc generation
Move all field descriptions to a single YAML file and generate version-
specific spec docs from that.  Use reflection to walk the config
structs for each version and omit fields which don't exist in a
particular spec version.

We don't do this at a JSON Schema level because e.g. Resource is
referenced in multiple places with different semantics and different
doc strings.

For now, keep the code out of the external API.  The plan is to add
functionality to support Butane docs generation, at which point
internal/doc/generate will move to config/doc.

Update the stabilization checklist to remove manual tweaking of
spec docs.