internal: move provider registrations out of platforms package
Move provider registrations next to the code that implements the provider,
as we do with stages.  This allows us to make the provider functions
private and only expose them through the registration.

To do this, we need to make the fields of the registration struct public,
which could allow the rest of Ignition to call directly through the
function pointers.  To avoid this, we wrap the registration struct when
exposing it to the rest of Ignition, ensuring that accessors are used.