internal/doc: reparent docs YAML root under a new component map
We'll eventually support multiple components, but for now, only support
the root component.

So far we've been treating the root of the doc as a slice of nodes, but we
already have node.Children, so that unnecessarily complicated things.
Define each component as a single node rather than a slice of them.  When
walking the root node, ignore its properties (such as description) and
skip directly to walking its children.

No functionality change intended.