internal/doc: use fully-qualified references to operator notes
Historically the Ignition spec docs have linked directly to the Markdown
file for the operator notes, which makes cross-references a little nicer
when viewing the spec docs directly on GitHub.  When Butane imports the
description text into its own spec docs, it then needs to translate the
link to a fully-qualified URL.  Now that spec docs are autogenerated, we
could continue converting the link, but it requires some extra transform
rules and doesn't seem worth it for the minor benefit we get.  Switch to
a fully-qualified URL in the Ignition spec docs.

For the same reason, explicitly mention Ignition in the link text.