Add optionally-installed grub2 code
This pairs with https://github.com/coreos/bootupd/pull/543
Basically we want bootupd to take care of installing this.  Add
the config here (where it more closely arguably belongs) so
that it can be optionally installed.

My proposal for at least the Fedora packaging of Ignition
is that we unconditionally depend on bootupd and install this,
because there's no one there using Ignition who doesn't want bootupd.