Add vsock modules into ramdisk
A review comment to #1696 suggested I see if I could eliminate a pair of
modules needed in the ramdisk to consume an ignition file over vsock.
In doing the testing, I experienced a false positive; unfortunately, I
have determined they are in fact needed and this PR asks to add them
again.

Signed-off-by: Brent Baude <bbaude@redhat.com>