tests/types: add configMax version to expand blackbox testing flexability
Add a field which can limit the max version of config generated from blackbox
testing. If left blank blackbox tests will act as they always have. The new
field simply limits the registration of tests to that version number. This
is for the odd case where we want to ensure functionality does not exist when
using older versions of configs.