Add initial tmt test and CI workflow
This commit introduces a new TMT tests and a GitHub Actions workflow to
run it.

TMT (Test Management Tool) is a tool for managing and executing tests
in a structured and efficient manner. It allows users to define test
plans, organize test cases, and execute them in a consistent manner. TMT is
particularly useful for testing software projects, as it provides a
framework for managing the entire testing process, from planning to
execution and reporting.

The workflow is set up to run on every push to the main branch and on
pull requests, providing continuous integration for the project.

The core test added is designed to verify the basic functionality of
the application using TMT, ensuring that it can successfully execute a
simple test case.

The smoke plan is defined in the `tests/tmt/plans/smoke.fmf` file and
is configure to run all tests with smoke tag.

The `tests/tmt/plans/main.fmf` file is the main plan that includes
common configurations and settings for the plans. Currently, it prepare
the test environment.

The workflow is defined in the `.github/workflows/tmt-tests.yml` file
and includes steps to set up the environment, install dependencies, and
run tests. It is allowed to execute on pull requests, pushes to the
main branch, and on demand. By default all tests are executed, but it
can be executed on demand with a TMT plan filter. As part of the
workflow, TMT will be executed inside a container.

JIRA: https://issues.redhat.com/browse/COS-2284

Signed-off-by: Tiago Bueno <tiago.bueno@gmail.com>